+++ src/main/java/com/iexec/core/configuration/PublicConfigurationService.java
@@ -16,10 +16,10 @@
 
 package com.iexec.core.configuration;
 
-import com.iexec.common.config.PublicConfiguration;
 import com.iexec.core.chain.ChainConfig;
 import com.iexec.core.chain.CredentialsService;
 import com.iexec.core.chain.adapter.BlockchainAdapterClientConfig;
+import com.iexec.core.config.PublicConfiguration;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.web3j.crypto.Hash;