+++ src/main/java/com/iexec/core/worker/WorkerController.java
@@ -16,11 +16,11 @@
 
 package com.iexec.core.worker;
 
-import com.iexec.common.config.PublicConfiguration;
 import com.iexec.common.config.WorkerModel;
 import com.iexec.commons.poco.security.Signature;
 import com.iexec.commons.poco.utils.BytesUtils;
 import com.iexec.commons.poco.utils.SignatureUtils;
+import com.iexec.core.config.PublicConfiguration;
 import com.iexec.core.configuration.PublicConfigurationService;
 import com.iexec.core.security.ChallengeService;
 import com.iexec.core.security.JwtTokenProvider;